Remove AppFormDate.vue
