+++ components/AppForm.vue
@@ -1,7 +1,21 @@
 <script lang="ts" setup>
-import { parseISO } from 'date-fns'
+import {
+  parseISO,
+  subDays,
+  subYears
+} from 'date-fns'
 
 const form = useFormStore()
+const now = new Date()
+const setToToday = () => {
+  form.date = formatISODate(now)
+}
+const setToYesterday = () => {
+  form.date = formatISODate(subDays(now, 1))
+}
+const setToOneYearAgo = () => {
+  form.date = formatISODate(subYears(now, 1))
+}
 const openTwitter = () => {
   const url = createTwitterSearchURL({
     user: form.user,
@@ -31,7 +45,37 @@ const handleEnter = (event: KeyboardEvent) => {
         autofocus
       />
     </el-form-item>
-    <AppFormDate />
+    <el-form-item label="日付">
+      <el-date-picker
+        v-model="form.date"
+        type="date"
+        value-format="YYYY-MM-DD"
+        clearable
+      />
+    </el-form-item>
+    <el-form-item>
+      <el-button
+        size="small"
+        :circle="false"
+        @click="setToToday"
+      >
+        今日
+      </el-button>
+      <el-button
+        size="small"
+        :circle="false"
+        @click="setToYesterday"
+      >
+        昨日
+      </el-button>
+      <el-button
+        size="small"
+        :circle="false"
+        @click="setToOneYearAgo"
+      >
+        1年前
+      </el-button>
+    </el-form-item>
     <AppFormKeyword />
     <AppFormIncludesRetweets />
     <el-form-item>
